ENH/BUG: Python server bug and user directory
The Python server was not setting the effective gid.  There was also a
bug in that traceback was not imported.  Added option so that the
python server chdirs to a user specific directory in order that a user
can save files which could be used as part of the test.