ENH: Added a Quiz model and time based papers
Added a model called Quiz which allows the admin to define a quiz on a
date which can be active or deactivated.  Users can take a quiz if one
is active and are given a fixed duration to complete the quiz.  They may
login as many times as needed to finish it but the total time is logged.
Javascript code updates a timer on the page.