ENH: Cleanup and adding error/comments for answers
Adding error and marks field to each answer.  Adding a new comment field
to the question paper and also a profile field for convenience.
Changing the views, templates and dump scripts to use the models rather
than Python code.  This cleans things up a lot more.  The user data
logged and printed is also way more comprehensive, paving the way for
easy online grading as well in the next phase of changes.