ENH: Creating a ServerPool for code checks
Changed the server so we use a pool of servers managed with a Queue of
available servers.  The XML/RPC client is also changed to handle
connection failures gracefully.  This is because XML/RPC cannot have
more than 2 connections at a given time, so if there is an error, we
wait for a random amount of time and try again.  This allows us to
handle fairly large loads nicely.