ENH: Fixing bash support, tests for code server.
This checkin fixes bash support.  In actuality the bash support lets one
test any runnable script/program that outputs results to stdout.  I've
also added a decent test suite for the code server that checks if it
functions correctly or not.  I've also updated the sample_questions to
work with the new bash support and added a reference bash script and the
testcode to go with it.