Function added to check java code
The function written by student is tested by this function.

The server side function calls the student function from
another java file containing the main function.

The main java file provides input to the student function
and checks for the expected output from the function.

If for certain set of input values, the output is as expec-
ted then the student code is graded correct.