Autochecks scilab function.
A method is created that checks the scilab function submitted by the
student.
The method removes the terminating commands in scilab.
If errors are present then it returns error,
else provides input to the function and checks for the
expected output.
If function returns correct output for all the inputs then
the function is graded correct, else throws the error message.