Removed for loop from the question template.
Once the code question is correct, user must see the message Correct
Output for two seconds and then url redirection.
But since this was written in a for loop, so
multiple requests almost simultaneously were sent by a single user.
This caused the server to go down. For now removed for loop.