Use a tornado based server for the pool server.
With the previous XMLRPC based server, an XMLRPC server would respond to
a request for an available port.  This does not work as the server can
only take about 2 simultaneous connections.  The server pool now uses a
HTTP server via tornado which works extremely well.  The django code
should not change at all as this is an internal change.  This change
should make the code server far more robust and work for a very large
number of simultaneous users.

The http server also has a simple status page to indicate the current
load.  This will not be correct on OSX due to limitations of the
multi-processing Queue implementation on OSX.