Views sometimes use cent percent CPU, fixed
After correct submission(POST) of code question, same question is shown for 2
seconds with a message "Correct Output".
After 2 seconds, the same correctly answered question is resubmitted(GET) to
the server.Since the question is already answered, it skips the question using
the skip method of answerpaper.
In skip method we have used cycle itertool, which loops in a cyclic
manner, never ending. So it is terminated when we get a question match
in an unanswered questions list with the submitted question. But the
question is already answered so we never get a match and loop runs
infinitely.
So used list instead of cycle.

Also, after correct answer, the user is to always get first question in
the answered list of question instead of next question after the
answered one.
So changed the completed_question method of answerpaper.