Answer saved again after correctly submitted, fixed.
Removed javascript that makes user wait for 2 seconds when the code
question is correct. All the other html elements are accessible by user
during the wait. This also caused the duplicate save during the wait, as
they can skip at that point and the answer is saved again.
Added a check that if question is already answered then do not save it.
This also resolves the monitor(use data) page problem of showing marks
obtained zero even when it is correct.
Removed skipped answers from the monitor page.