Forgot Password and Reset Issue Resolved
The problem was:
We have included yaksh urls to the project urls with the namespace
"yaksh".
So whenever we call the url name i.e "yaksh:<url>" a reverse match is
made.
But for Forgot Password and Change Password we are using django's
in-built functionality, where the reverse url is hardcoded.
So the reverse match fails as the namespace is not specified!
To resolve this, created a urls_password_reset URL dispatcher file and
included to the project urls.