Fix #201.
The issue is that when the ouput directory is set to
`tempfile.gettempdir()` the user nobody cannot stat the created
directory on OS X.  The code server therefore is unable to see the
directory resulting in problems.  Instead we create the directory inside
the application as it was done earlier.