Added migrations folder to the app.
This required because when we do migration for the first time, initial
migrations will be created only if migrations folder is present in the
installed app. Otherwise, running makemigrations will not detect the
initial migrations for that app.
Though we can exclusively specify the app name as the argument to the
makemigrations command.
But did this so that by default apply migrations.

Note: Removed migrations from .gitignore