Removed a function call to datetime now in default quiz start time.
Every time when we run makemigrations, model changes are detected.
"Alter field start_date_time on quiz" was always detected, even if you
have not mode any changes to the models!
This was happening because of default time value was datetime.now().
So every time when you ran migrations current datetime was set which will
always change.
Now removed parenthesis(), so that now function is not called.
And no unnecessary model changes will be detected