Completely rewrite the server pool.
It is now much simpler.  There are no xmlrpc servers anymore instead the
tornado server takes a post request to submit a job asynchronously and
the results are added to a shared dictionary.  A get request can be used
to check the status of a submitted job.  This allows the submission and
checking of code to be completely async and will make the application a
lot more scalable.