Changes related to front-end
Removed snippet append in the check view as snippet is not posted in
request.
Added an overlay when an user submits a code, with a status text on it.
This is to block user from triggering any other event when JS is
running. Overlay disappears when JS complete its execution.
On time out a request is posted via JS, it receives a JSON response but
cannot display user the error as time is over. So in such case, the
django itself handles the result and does not return JSONv response.