Added 8 seconds sleep in selenium test for now.
As implicit wait is not functioning as desired.
Added sleep for now. Since new exam interface has AJAX request response,
so interface sets a blank div on top of all divs to prevent further
events till a response is received. So selenium does not get the
required HTML element at that period.
And 8 seconds because for infinite loop server responds after 4 seconds.