Use bulk create for creating the profiles
For  creating users, did not add bulk create as we need the created user
for creating profile and to add in the course.
So the course and profile is linked with the user. And bulk create does
not return the primary key, so becomes complex to create profile or add
user to the course.